fix: simplify pass — bugs, resource leaks, deduplication
- Fix CorrectionPopover sending source:"user" (rejected by backend) → "ui"
- Fix defer inside loops in learned.go (resource leak on HTTP bodies)
- Add sanitizeParam on projectName in feedback_loop_config handlers
- Hoist regexp.MustCompile to package-level var in extraction.go
- Unify feature flag constant (remove duplicate feedbackLoopFeatureFlag)
- Extract CORRECTION_TYPE_LABELS to shared corrections.ts
- Extract parseOwnerRepo to shared lib/github-utils.ts
