+++ components/frontend/src/services/api/corrections.ts
@@ -8,6 +8,14 @@ import { apiClient } from './client';
 /** Correction types matching runner's CORRECTION_TYPES */
 export type CorrectionType = 'incomplete' | 'incorrect' | 'out_of_scope' | 'style';
 
+/** Human-readable labels for correction types */
+export const CORRECTION_TYPE_LABELS: Record<CorrectionType, string> = {
+  incomplete: "Incomplete",
+  incorrect: "Incorrect",
+  out_of_scope: "Out of Scope",
+  style: "Style",
+};
+
 /** Request payload for submitting a correction */
 export type SubmitCorrectionRequest = {
   correction_type: CorrectionType;
@@ -15,7 +23,7 @@ export type SubmitCorrectionRequest = {
   session_name: string;
   message_id: string;
   message_content?: string;
-  source: 'user';
+  source: 'ui';
 };
 
 /** Response from the corrections endpoint */
